// frontend/src/config/contracts.generated.ts
import type { Address } from 'viem'

// This file is auto-generated by `npm run contracts:sync`.
// Do not edit it manually.

export const fluxContractNames = ["FluxBuybackExecutor","FluxMultiPoolManager","MockUSDC","MockUSDT","MockWBTC","FluxPoolFactory","FluxRevenueDistributor","FluxSignedOrderSettlement","FluxSwapFactory","FluxSwapRouter","FluxSwapTreasury","FluxToken","MockWETH"] as const

export type FluxContractName = (typeof fluxContractNames)[number]
export type FluxContractAddressMap = Partial<Record<FluxContractName, Address>>

export const generatedContractAddressesByChain = {
  31337: {
    FluxBuybackExecutor: '0x610178dA211FEF7D417bC0e6FeD39F05609AD788' as Address,
    FluxMultiPoolManager: '0xa513E6E4b8f2a923D98304ec87F64353C4D5C853' as Address,
    MockUSDC: '0xDc64a140Aa3E981100a9becA4E685f962f0cF6C9' as Address,
    MockUSDT: '0x5FC8d32690cc91D4c39d9d3abcBD16989F875707' as Address,
    MockWBTC: '0x0165878A594ca255338adfa4d48449f69242Eb8F' as Address,
    FluxPoolFactory: '0xB7f8BC63BbcaD18155201308C8f3540b07f84F5e' as Address,
    FluxRevenueDistributor: '0x0DCd1Bf9A1b36cE34237eEaFef220932846BCD82' as Address,
    FluxSignedOrderSettlement: '0xA51c1fc2f0D1a1b8494Ed1FE312d7C3a78Ed91C0' as Address,
    FluxSwapFactory: '0x5FbDB2315678afecb367f032d93F642f64180aa3' as Address,
    FluxSwapRouter: '0x2279B7A0a67DB372996a5FaB50D91eAA73d2eBe6' as Address,
    FluxSwapTreasury: '0xe7f1725E7734CE288F8367e1Bb143E90bb3F0512' as Address,
    FluxToken: '0x9fE46736679d2D9a65F0992F2272dE9f3c7fa6e0' as Address,
    MockWETH: '0xCf7Ed3AccA5a467e9e704C703E8D87F634fB0Fc9' as Address,
  },
  11155111: {
    FluxBuybackExecutor: '0xBF2adc65825217329813BE599c6785bA4f07c409' as Address,
    FluxMultiPoolManager: '0x8faf75D83DA098Eb9795D251707201eB80Da3F05' as Address,
    MockUSDC: '0xA4049489D8689A55154c35f9809B1382B0972A79' as Address,
    MockUSDT: '0xf4fc2De27781B6401609732e0aa7f7c354BdE388' as Address,
    MockWBTC: '0xC97de7b464a4dB1910c16ce7835b918B4440C42C' as Address,
    FluxPoolFactory: '0x6E6B4E527ea8e5c8a3c4789A502182656A291E96' as Address,
    FluxRevenueDistributor: '0x119261fD7CF64A336312A41BF5da682241F2849b' as Address,
    FluxSignedOrderSettlement: '0x53fCB3a81398F2f7c429A3471E44D411e360594a' as Address,
    FluxSwapFactory: '0x4081e489309CF890838fa6Ce7b563Bc198d5B715' as Address,
    FluxSwapRouter: '0x85ba499C5E8DD41C2BeF7571BE68AaD4c82Bf6DF' as Address,
    FluxSwapTreasury: '0xc993b497f1E7f997130B8BD1Db8EFb5169faEf48' as Address,
    FluxToken: '0x5171c0FEc2F9C4413a9AAe39F3D3739Aa30b1110' as Address,
    MockWETH: '0xf881723bEF047DA1CD494795a797835241D3be29' as Address,
  },
} as const satisfies Record<number, FluxContractAddressMap>